Add GeneratorFictitious model
Modified some model of generators : the omega and omegaRef are only needed in synchronous generators.
Beware, this generator is not supposed to be linked to the OmegaRef, as this generator has no way to be synchronized.
TODO: find a way to no instanciate a OmegaRef when using the Fictitious generator

Signed-off-by: BAUDRIER Dimitri <dimitri.baudrier@rte-france.com>